FastApi_Client/requirements.txt
anyio==3.6.2
certifi==2023.5.7
fastapi==0.95.1
grpcio==1.54.0
grpcio-tools==1.54.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
protobuf==4.22.4
pydantic==1.10.7
pytest==7.3.1
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
